Upload options

+++ aip2b2.rb
@@ -1,6 +1,13 @@
 #!/usr/bin/env ruby
 require "pathname"
 require 'json'
+require 'optparse'
+
+ARGV.options do |opts|
+  opts.on("-d", "--dry-run")  { $dryrun = '--dryRun ' }
+  opts.on("-p", "--path=val", String)  { |val| $b2path = val }
+  opts.parse!
+end
 
 # Set up methods
 def premisreport(actiontype,outcome)
@@ -41,27 +48,41 @@ class String
   end
 end
 
- ARGV.each do |input_AIP|
-   @target_path = Pathname.new(input_AIP)
-   # Test for directory
-   if ! @target_path.directory?
-     puts "Input must be a directory! Exiting.".red && exit
-   end
-   @targetdir = File.dirname(input_AIP)
-   packagename = File.basename(@target_path)
-   b2_target = 'b2://INSERT-PATH-HERE' + packagename
-   logfile = @target_path + 'data' + 'logs' + "#{packagename}.log"
-   @premis_structure = JSON.parse(File.read(logfile))
+# Check for b2 path
+if $b2path.nil?
+  puts "Please enter a B2 path using the -p flag.".red
+  exit
+end
+
+ARGV.each do |input_AIP|
+  @target_path = Pathname.new(input_AIP)
+  # Test for directory
+  if ! @target_path.directory?
+    puts "Input must be a directory! Exiting.".red && exit
+  end
+  packagename = File.basename(@target_path)
+  b2_target = $b2path + '/' + packagename
+  logfile = @target_path + 'data' + 'logs' + "#{packagename}.log"
+  @premis_structure = JSON.parse(File.read(logfile))
+  if $dryrun.nil?
    $command = 'b2 sync ' + '"' + @target_path.to_s + '" ' + '"' + b2_target + '"'
-   if system($command)
+  else
+    $command = 'b2 sync ' + $dryrun + '"' + @target_path.to_s + '" ' + '"' + b2_target + '"'
+  end
+
+  if system($command)
     puts "SUCCESS!".green
     premisreport('replication','pass')
     puts @premis_structure
-    File.open("#{@target_path.to_s}/#{packagename}_#{Time.now}_premis.log",'w') {|file| file.write(@premis_structure.to_json)}
-    system($command)
-    outcomereport('premis')
-   else
-    premisreport('replication','fail')
-    outcomereport('premis')
-   end
- end
+    if $dryrun.nil?
+      File.open("#{@target_path.to_s}/#{packagename}_#{Time.now}_premis.log",'w') {|file| file.write(@premis_structure.to_json)}
+      system($command)
+      outcomereport('premis')
+    end
+  else
+    if $dryrun.nil?
+      premisreport('replication','fail')
+      outcomereport('premis')
+    end
+  end
+end
